Drop old and unused files, remove dummy from module names
