+++ app/index.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const clientApp = require('../clientapp');
-const api = require('./dummy-api');
+const api = require('./api');
 const config = require('./lib/config');
 const nunjucks = require('nunjucks');
 const persona = require('express-persona');
