--- bin/db-setup
@@ -1,42 +1,0 @@
-#!/usr/bin/env node
-
-var db = require('../app/lib/db');
-var async = require('async');
-
-async.series([
-
-  function (cb) {
-    console.log('Deleting all');
-    db.deleteAll(function (err) {
-      if (err) return cb(err);
-      console.log('\t--> Ok');
-      cb();
-    });
-  },
-
-  function (cb) {
-    var q = "CREATE CONSTRAINT ON (user:User) ASSERT user.email IS UNIQUE";
-    console.log(q);
-    db.query(q, {}, function (err) {
-      if (err && err.message.indexOf('Already constrained') === -1) return cb(err);
-      else console.log('\t--> Ok');
-      cb();
-    });
-  },
-
-  function (cb) {
-    var q = "MERGE (p:Pathway {name: 'Demo Pathway'})" +
-      " FOREACH (i in range(0, 3)|" +
-      " CREATE UNIQUE p-[:contains]->(:Requirement {x: i, y: i})-[:references]->(:BadgeClass {name: 'Badge ' + i}))" +
-      " RETURN p";
-    console.log(q) 
-    db.query(q, function (err, results) {
-      if (err) return cb(err);
-      console.log('\t--> Ok');
-      cb();
-    });
-  }
-], function (err) {
-  if (err) throw err;
-  console.log('\t--> Done!');
-});
